Removed o_ prefix from o_properties property
